# Settings being cached now, huge performance boost