--- a/ExilenceClient/src/app/shared/providers/electron.service.ts
+++ b/ExilenceClient/src/app/shared/providers/electron.service.ts
@@ -67,7 +67,6 @@ export class ElectronService {
   }
 
   private sendLogToServer(log: string) {
-
     const accountName = this.settings.get('account').accountName;
     const settings = this.settings.getAll();
     const stringSettings = JSON.stringify(settings);