--- a/ExilenceClient/src/app/shared/providers/income.service.ts
+++ b/ExilenceClient/src/app/shared/providers/income.service.ts
@@ -83,13 +83,15 @@ export class IncomeService implements OnDestroy {
 
   Snapshot() {
     const oneDayAgo = (Date.now() - (24 * 60 * 60 * 1000));
-
     const twoWeeksAgo = (Date.now() - (1 * 60 * 60 * 24 * 14 * 1000));
 
-    this.netWorthHistory = this.settingsService.get('networth');
-    const selectedStashtabs = this.settingsService.get('selectedStashTabs');
 
-    this.sessionIdValid = this.settingsService.get('account.sessionIdValid');
+    const settings = this.settingsService.getAll();
+
+    this.netWorthHistory = settings.networth;
+    const selectedStashtabs = settings.selectedStashTabs;
+
+    this.sessionIdValid = settings.account.sessionIdValid;
     if (
       this.netWorthHistory.lastSnapshot < (Date.now() - this.threeMinutes) &&
       this.localPlayer !== undefined &&
@@ -221,7 +223,9 @@ export class IncomeService implements OnDestroy {
     this.totalNetWorthItems = [];
     this.totalNetWorth = 0;
 
-    const itemValueTresholdSetting = this.settingsService.get('itemValueTreshold');
+    const settings = this.settingsService.getAll();
+
+    const itemValueTresholdSetting = settings.itemValueTreshold;
     if (itemValueTresholdSetting !== undefined) {
       this.itemValueTreshold = itemValueTresholdSetting;
     } else {
@@ -229,7 +233,7 @@ export class IncomeService implements OnDestroy {
       this.settingsService.set('itemValueTreshold', 1);
     }
 
-    const characterPricing = this.settingsService.get('characterPricing');
+    const characterPricing = settings.characterPricing;
     if (characterPricing !== undefined) {
       this.characterPricing = characterPricing;
     } else {
@@ -237,7 +241,7 @@ export class IncomeService implements OnDestroy {
       this.settingsService.set('characterPricing', false);
     }
 
-    const selectedStashTabs: any[] = this.settingsService.get('selectedStashTabs');
+    const selectedStashTabs: any[] = settings.selectedStashTabs;
 
     let mapTab;
     if (selectedStashTabs !== undefined) {