Each time a setting had to be loaded, whole file was read in and the setting was searched from inside. Now settings are cached in memory and it boost performance 10x times.

Also optimized a for cycle that was writing to file each cycle, which was very costly because it read the whole file in again and then had to write it.

**Biggest loading time improvements:**

1. Pressing settings button previously took me ~25 seconds. Now it takes 1.5 seconds
2. Joining group took me ~20 seconds, now takes 9 seconds, where scripting part reduced from 11.5 seconds to 1.5 seconds, but rest time is behind getting stash etc from path of exile, so that can't be improved further.


My settings file is 7mb large and I don't play too much. Active person will have it much larger. The bigger the file, the slower it was.