@viktorgullmark 2019-01-20
Great work! Will take a closer look at this tomorrow. If everything works as expected, I'll merge this before the next release.